Support for color, italic and underline in EIA608 subtitles
Style information was not parsed before, but now I added support
for various styling described in the EIA608 specification.
Positioning (Row, Column and Tab handling) is not implemented yet.

The standard describes 2 fields (of the analog transmission), 2 channels
(in a single closed caption stream) and other requirements that are quite
deprecated and does not really make sense with digital transmission.
A second channel (other language) should be an independent track and not
mixed into the first one any more. So the current implementation of the
608Parser, Control Codes and 608TrackRenderer only handle a single
channel and does not differentiate odd and even fields (of the analog
transmission).  

Change-Id: Ibf759cd1420c93f617b612effd123ff4ae4c17a5